refactor(dsh-mcp-manager): 客户端分层 core/float/settings 落位
Summary of changes:
- src/client/ 按 v3 目标树分层:core/(state/session/dom/api/i18n/constants)、
  float/(float/panel/servers/quick-add)、settings/(settings-card.tsx);入口
  index.ts 原地保留,import 路径随分层同步
- core/dom.ts 保留 el(DOM 创建),core/api.ts 收 api()(HTTP 请求,拆出自
  dom.ts,C14 204 备忘注释随迁);core/i18n.ts 收 tStatus 渲染期文案求值
  (拆出自 locales.ts),字典与 McpLocaleKey 类型留根
- shared/client/* 相对路径二级文件加深一级(4 层→5 层),一级文件维持 4 层
- docs/architecture-redesign.md 补《阶段 7 客户端分层落地映射表》与客户端
  目录不变式
- 零行为变更:纯搬移 + 拆分,构建/verify-dir-imports/测试全绿

Design decisions and rationale:
- 客户端不受 D10 门面约束(src/client/ 从 verify-dir-imports from 侧整体
  豁免,index.ts 为 build-client 契约锚点),分层仅物理归位,不建目录门面
- C 类修复另行 commit,refactor 与行为修复分离可独立验证(阶段不变式 3)

Related issue/ticket references: #664
