+++ packages/dsh-mcp-manager/src/client/settings/settings-card.tsx
@@ -7,9 +7,9 @@
  */
 
 import * as React from "react";
-import { API } from "./constants.ts";
-import { api } from "./dom.ts";
-import { t } from "../../../../shared/client/i18n.js";
+import { API } from "../core/constants.ts";
+import { api } from "../core/api.ts";
+import { t } from "../../../../../shared/client/i18n.js";
 
 /**
  * 设置页插件卡（settings.plugin.item）：浮窗位置 / 偏移编辑区。
